api/requirements.txt
fastapi==0.118.3
uvicorn==0.37.0
requests==2.32.5
psycopg2-binary==2.9.11
python-dotenv==1.1.1